test: Move 'simple_common_sim' fixture to conftest
